backend/requirements.txt
tensorflow
numpy
flask
flask_cors
google-generativeai
dotenv
gunicorn